Refactor: merge util_refresh & util_cache methods into an impl for JwtKeyCache
