+++ src/oauth2/jwk/mod.rs
@@ -48,5 +48,3 @@ pub mod jwk;
 mod task;
 
 mod util;
-mod util_cache;
-mod util_refresh;
